raspberry_pi/requirements.txt
numpy>=1.24
opencv-python>=4.8
pyserial>=3.5
PyYAML>=6.0
pytest>=7.4
